+++ src/app/compare/compare.service.ts
@@ -3,18 +3,11 @@ import { Subject } from 'rxjs/Subject';
 
 export const CompareService = {
   dialogOpen: false,
-  compareType: '',
-  compareItem$: new Subject<{
-    item: DimItem;
+  compareItems$: new Subject<{
+    items: DimItem[];
     dupes: boolean;
   }>(),
-  compareMultiItem$: new Subject<{
-    items: DimItem[];
-  }>(),
-  addItemToCompare(item: DimItem, dupes = false) {
-    this.compareItem$.next({ item, dupes });
-  },
-  addMultiItemsToCompare(items: DimItem[]) {
-    this.compareMultiItem$.next({ items });
+  addItemsToCompare(items: DimItem[], dupes = false) {
+    this.compareItems$.next({ items, dupes });
   }
 };